+++ src/main/java/truckmanagementproject/web/controllers/users/ManagerController.java
@@ -6,11 +6,14 @@ import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
+import truckmanagementproject.services.services.auth.AuthService;
 import truckmanagementproject.services.services.managers.ManagerService;
 import truckmanagementproject.services.models.managers.AddManagerServiceModel;
+import truckmanagementproject.web.models.auth.LoginUserViewModel;
 import truckmanagementproject.web.models.managers.AddManagerModel;
 import truckmanagementproject.web.models.managers.ManagerViewModel;
 
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -26,16 +29,21 @@ public class ManagerController {
 
     private final ModelMapper mapper;
     private final ManagerService managerService;
+    private final AuthService authService;
 
     @Autowired
-    public ManagerController(ModelMapper mapper, ManagerService managerService) {
+    public ManagerController(ModelMapper mapper, ManagerService managerService, AuthService authService) {
         this.mapper = mapper;
         this.managerService = managerService;
+        this.authService = authService;
     }
 
-    //TODO -> ONLY ACCESSIBLE FROM ADMIN
     @GetMapping("/add")
-    public String getAddManagerForm(@ModelAttribute("model") AddManagerModel model) {
+    public String getAddManagerForm(@ModelAttribute("model") AddManagerModel model, HttpSession session) throws Exception {
+        LoginUserViewModel user = (LoginUserViewModel) session.getAttribute("user");
+        if (!authService.isUserAdmin(user)) {
+            throw new Exception("Unauthorized user");
+        }
         return "managers/add";
     }
 
@@ -55,9 +63,12 @@ public class ManagerController {
         }
     }
 
-    //TODO -> ONLY ACCESSIBLE FROM ADMIN + MANAGER
     @GetMapping("/all")
-    public ModelAndView getAllManagers(ModelAndView modelAndView) {
+    public ModelAndView getAllManagers(ModelAndView modelAndView, HttpSession session) throws Exception {
+        LoginUserViewModel user = (LoginUserViewModel) session.getAttribute("user");
+        if (!authService.isUserAdmin(user) && !authService.isUserAdmin(user)) {
+            throw new Exception("Unauthorized user");
+        }
         List<ManagerViewModel> managers = managerService.getAllManagers()
                 .stream()
                 .map(manager -> mapper.map(manager, ManagerViewModel.class))
@@ -68,9 +79,12 @@ public class ManagerController {
         return modelAndView;
     }
 
-    //TODO -> ONLY ACCESSIBLE FROM ADMIN
     @GetMapping("/remove/{id}")
-    public ModelAndView removeManager(@PathVariable String id) {
+    public ModelAndView removeManager(@PathVariable String id, HttpSession session) throws Exception {
+        LoginUserViewModel user = (LoginUserViewModel) session.getAttribute("user");
+        if (!authService.isUserAdmin(user)) {
+            throw new Exception("Unauthorized user");
+        }
         managerService.removeManager(id);
         return new ModelAndView("redirect:/managers/all");
     }
